jconf package API change! Read below
It was deemed necessary to make a minor change in the jconf API.
Ratinale: Config parsing shouldn't panic due to spelling errors.

The previous code would panic during parsing if you spelled a JSON
key for a SubConfig wrong, - since the correct key would then not be there
and ParseInto() would be called on a nil value unless everybody checked
everywhere for nil SubConfigs.
Always returning an error in that case was not an options since some
SubConfigs could be optional.
So instead SubConfig was changed to an interface having both Mandatory
and Optional variants.

The code modifications needed are minor:
* Change *jconf.SubConfig to *jconf.MandatorySubconfig in your struct
  definitions. (or the OptionalSubconfig)
* Change *jconf.SubConfig to the interface jconf.SubConfig in any
  parameter parsing.

Overall this is also a better design, since futher special subconfigs
could be added and you can change which one you use easily.