Cleanup metric API and satisfy go linter.
The metric API had a lot of exported API which should really be internal.
For the part of the API which was actually supposed to be used by applications there
is no change, but various internal types and functions has been made not exported.
A few missing functions has been added.